Move Storage to core
